Update potion.py
./kyu_6/potion_class_101/potion.py:63 in private method `__calc_rgb`:
        D205: 1 blank line required between summary line and description (found 0)

Multi-line docstrings consist of a summary line just like a one-line
docstring, followed by a blank line, followed by a more elaborate
description. The summary line may be used by automatic indexing tools;
it is important that it fits on one line and is separated from the
rest of the docstring by a blank line.