Update test_replace_with_alphabet_position.py
./kyu_6/replace_with_alphabet_position/test_replace_with_alphabet_position.py:79 in public method `test_alphabet_position_edge_cases`:
        D205: 1 blank line required between summary line and description (found 0)

Multi-line docstrings consist of a summary line just like a one-line
docstring, followed by a blank line, followed by a more elaborate
description. The summary line may be used by automatic indexing tools;
it is important that it fits on one line and is separated from the
rest of the docstring by a blank line.

./kyu_6/replace_with_alphabet_position/test_replace_with_alphabet_position.py:79 in public method `test_alphabet_position_edge_cases`:
        D400: First line should end with a period (not ',')

The [first line of a] docstring is a phrase ending in a period.

./kyu_6/replace_with_alphabet_position/test_replace_with_alphabet_position.py:79 in public method `test_alphabet_position_edge_cases`:
        D415: First line should end with a period, question mark, or exclamation point (not ',')

The [first line of a] docstring is a phrase ending in a period,
question mark, or exclamation point